Added flag 'make_gsl_orog' so the user can create extra 'orog' files or not

+++ ush/fv3gfs_driver_grid.sh
@@ -16,7 +16,8 @@
 #   2) 'oro' files containing land mask, terrain and gravity
 #      wave drag fields.
 #   3) 'oro' files ('oro_data_ls' and 'oro_data_ss') specific to
-#      the GSL drag suite physics parameterization
+#      the GSL drag suite physics parameterization (only if
+#      flag make_gsl_orog = true)
 #   4) surface climo fields, such as soil type, vegetation
 #      greenness and albedo.
 #
@@ -174,12 +175,16 @@ if [ $gtype = uniform ] || [ $gtype = stretch ] || [ $gtype = nest ];  then
 
   if [ $machine = WCOSS_C ]; then
     touch $TEMP_DIR/orog.file1
-    touch $TEMP_DIR/orog_gsl.file1
+    if [ $make_gsl_orog = true ]; then
+      touch $TEMP_DIR/orog_gsl.file1
+    fi
     tile=1
-    export halo_tmp="-999"  # no halo
     while [ $tile -le $ntiles ]; do
       echo "$script_dir/fv3gfs_make_orog.sh $res $tile $grid_dir $orog_dir $script_dir $topo " >>$TEMP_DIR/orog.file1
-      echo $script_dir/fv3gfs_make_orog_gsl.sh $res $tile $halo_tmp $grid_dir $orog_dir $topo_am >>$TEMP_DIR/orog_gsl.file1
+      if [ $make_gsl_orog = true ]; then
+        export halo_tmp="-999"  # no halo
+        echo $script_dir/fv3gfs_make_orog_gsl.sh $res $tile $halo_tmp $grid_dir $orog_dir $topo_am >>$TEMP_DIR/orog_gsl.file1
+      fi
       tile=$(( $tile + 1 ))
     done
     aprun -j 1 -n 4 -N 4 -d 6 -cc depth cfp $TEMP_DIR/orog.file1
@@ -188,12 +193,14 @@ if [ $gtype = uniform ] || [ $gtype = stretch ] || [ $gtype = nest ];  then
       exit $err
     fi
     rm $TEMP_DIR/orog.file1
-    aprun -j 1 -n 4 -N 4 -d 6 -cc depth cfp $TEMP_DIR/orog_gsl.file1
-    err=$?
-    if [ $err != 0 ]; then
-      exit $err
+    if [ $make_gsl_orog = true ]; then
+      aprun -j 1 -n 4 -N 4 -d 6 -cc depth cfp $TEMP_DIR/orog_gsl.file1
+      err=$?
+      if [ $err != 0 ]; then
+        exit $err
+      fi
+      rm $TEMP_DIR/orog_gsl.file1
     fi
-    rm $TEMP_DIR/orog_gsl.file1
   else
     tile=1
     while [ $tile -le $ntiles ]; do
@@ -207,16 +214,18 @@ if [ $gtype = uniform ] || [ $gtype = stretch ] || [ $gtype = nest ];  then
       if [ $err != 0 ]; then
         exit $err
       fi
-      set +x
-      echo
-      echo "............ Execute fv3gfs_make_orog_gsl.sh for tile $tile .................."
-      echo 
-      set -x
-      export halo_tmp="-999"  # no halo
-      $script_dir/fv3gfs_make_orog_gsl.sh $res $tile $halo_tmp $grid_dir $orog_dir $topo_am
-      err=$?
-      if [ $err != 0 ]; then
-        exit $err
+      if [ $make_gsl_orog = true ]; then
+        set +x
+        echo
+        echo "............ Execute fv3gfs_make_orog_gsl.sh for tile $tile .................."
+        echo 
+        set -x
+        export halo_tmp="-999"  # no halo
+        $script_dir/fv3gfs_make_orog_gsl.sh $res $tile $halo_tmp $grid_dir $orog_dir $topo_am
+        err=$?
+        if [ $err != 0 ]; then
+          exit $err
+        fi
       fi
       tile=$(( $tile + 1 ))
     done
@@ -258,8 +267,10 @@ if [ $gtype = uniform ] || [ $gtype = stretch ] || [ $gtype = nest ];  then
   tile=1
   while [ $tile -le $ntiles ]; do
     cp $filter_dir/oro.C${res}.tile${tile}.nc $out_dir/C${res}_oro_data.tile${tile}.nc
-    cp $orog_dir/C${res}_oro_data_*.tile${tile}*.nc $out_dir/  # gsl drag suite oro_data files
     cp $grid_dir/C${res}_grid.tile${tile}.nc  $out_dir/C${res}_grid.tile${tile}.nc
+    if [ $make_gsl_orog = true ]; then
+      cp $orog_dir/C${res}_oro_data_*.tile${tile}*.nc $out_dir/  # gsl drag suite oro_data files
+    fi
     tile=`expr $tile + 1 `
   done
 
@@ -477,31 +488,33 @@ elif [ $gtype = regional_gfdl ] || [ $gtype = regional_esg ]; then
 
 
 #----------------------------------------------------------------------------------
-# Now that C${res}_grid.tile${tile}.halo0.nc has been created, we can
-# use it to generate gsl drag suite oro_data files, which is only generated
-# for halo0
+# Now that C${res}_grid.tile${tile}.halo0.nc has been created, we can use it
+# to generate gsl drag suite oro_data files, which are generated only for halo0
+# Note:  This is carried out only if $make_gsl_orog = true
 #----------------------------------------------------------------------------------
 
-  export halo_tmp="0"
-  ln -sf $out_dir/C${res}_grid.tile${tile}.halo0.nc $grid_dir/
-  if [ $machine = WCOSS_C ]; then
-    echo $script_dir/fv3gfs_make_orog_gsl.sh $res $tile $halo_tmp $grid_dir $orog_dir $topo_am >>$TEMP_DIR/orog_gsl.file1
-    aprun -j 1 -n 4 -N 4 -d 6 -cc depth cfp $TEMP_DIR/orog_gsl.file1
-    err=$?
-    rm $TEMP_DIR/orog_gsl.file1
-  else
-    set +x 
-    echo
-    echo "............ Execute fv3gfs_make_orog_gsl.sh for tile $tile .................."
-    echo
-    set -x
-    $script_dir/fv3gfs_make_orog_gsl.sh $res $tile $halo_tmp $grid_dir $orog_dir $topo_am
-    err=$?
-    if [ $err != 0 ]; then
-      exit $err
+  if [ $make_gsl_orog = true ]; then
+    export halo_tmp="0"
+    ln -sf $out_dir/C${res}_grid.tile${tile}.halo0.nc $grid_dir/
+    if [ $machine = WCOSS_C ]; then
+      echo $script_dir/fv3gfs_make_orog_gsl.sh $res $tile $halo_tmp $grid_dir $orog_dir $topo_am >>$TEMP_DIR/orog_gsl.file1
+      aprun -j 1 -n 4 -N 4 -d 6 -cc depth cfp $TEMP_DIR/orog_gsl.file1
+      err=$?
+      rm $TEMP_DIR/orog_gsl.file1
+    else
+      set +x 
+      echo
+      echo "............ Execute fv3gfs_make_orog_gsl.sh for tile $tile .................."
+      echo
+      set -x
+      $script_dir/fv3gfs_make_orog_gsl.sh $res $tile $halo_tmp $grid_dir $orog_dir $topo_am
+      err=$?
+      if [ $err != 0 ]; then
+        exit $err
+      fi
     fi
+    cp $orog_dir/C${res}_oro_data_*.tile${tile}*.nc $out_dir/  # gsl drag suite oro_data files
   fi
-  cp $orog_dir/C${res}_oro_data_*.tile${tile}*.nc $out_dir/  # gsl drag suite oro_data files
 
   echo "Grid and orography files are now prepared for regional grid"
